feat(pratofeito-backend): Default monorepo structure with pom module for libs and core-api
This is our initial monorepo configuration. It's important to understand the main structure and the possibilities of decoupling - and that's one reason why we are using Kotlin. It's easier to define organization and encapsulation in a way that in the future we will have different clients for the core-api and domain model.